Move common types and common utils to embabel-common modules
