+++ embabel-agent-api/src/main/kotlin/com/embabel/agent/core/AgentProcess.kt
@@ -18,7 +18,7 @@ package com.embabel.agent.core
 import com.embabel.common.core.types.HasInfoString
 import com.embabel.common.core.types.Timed
 import com.embabel.common.core.types.Timestamped
-import com.embabel.common.core.util.ComputerSaysNoSerializer
+import com.embabel.common.util.ComputerSaysNoSerializer
 import com.embabel.plan.WorldState
 import com.fasterxml.jackson.databind.annotation.JsonSerialize
 import java.time.Duration
