--- embabel-agent-api/src/main/kotlin/com/embabel/common/core/util/ComputerSaysNoSerializer.kt
@@ -1,38 +1,0 @@
-/*
- * Copyright 2024-2025 Embabel Software, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.embabel.common.core.util
-
-import com.fasterxml.jackson.core.JsonGenerator
-import com.fasterxml.jackson.databind.JsonSerializer
-import com.fasterxml.jackson.databind.SerializerProvider
-
-/**
- * Use @JsonSerialize(using = ComputerSaysNoSerializer.class) to prevent serialization of
- * a class. This can be important to safeguard domain objects from being serialized and
- * sent to the client.
- *
- * https://www.youtube.com/watch?v=x0YGZPycMEU&ab_channel=MattLucasandDavidWalliams
- *
- * @param <T>
-</T> */
-class ComputerSaysNoSerializer<T> : JsonSerializer<T>() {
-
-    override fun serialize(value: T, gen: JsonGenerator, serializers: SerializerProvider) {
-        throw IllegalArgumentException(
-            "Computer says no to serializing objects of type ${value!!::class.java.name} as it is internal"
-        )
-    }
-}
