--- embabel-agent-api/src/main/kotlin/com/embabel/common/core/util/NameUtils.kt
@@ -1,34 +1,0 @@
-/*
- * Copyright 2024-2025 Embabel Software, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.embabel.common.core.util
-
-// TODO move to common and merge with NamingUtils
-object NameUtils {
-
-    fun beanMethodToPropertyName(name: String): String {
-        return if (name.startsWith("get") || name.startsWith("set")) {
-            name.substring(3).replaceFirstChar { it.lowercase() }
-        } else if (name.startsWith("is")) {
-            name.substring(2).replaceFirstChar { it.lowercase() }
-        } else {
-            name
-        }
-    }
-
-    fun isBeanMethod(name: String): Boolean {
-        return name.startsWith("get") || name.startsWith("set") || name.startsWith("is")
-    }
-}
